refactor: modify validation middleware
